Move requirement logic inside report function
