+++ internal/runbits/cves/cves.go
@@ -14,9 +14,11 @@ import (
 	"github.com/ActiveState/cli/internal/primer"
 	"github.com/ActiveState/cli/internal/rtutils/ptr"
 	"github.com/ActiveState/cli/pkg/buildplan"
+	"github.com/ActiveState/cli/pkg/buildscript"
 	vulnModel "github.com/ActiveState/cli/pkg/platform/api/vulnerabilities/model"
 	"github.com/ActiveState/cli/pkg/platform/api/vulnerabilities/request"
 	"github.com/ActiveState/cli/pkg/platform/model"
+	"github.com/ActiveState/cli/pkg/platform/model/buildplanner"
 )
 
 func init() {
@@ -39,8 +41,8 @@ func NewCveReport(prime primeable) *CveReport {
 	return &CveReport{prime}
 }
 
-func (c *CveReport) Report(newBuildPlan *buildplan.BuildPlan, oldBuildPlan *buildplan.BuildPlan, names ...string) error {
-	changeset := newBuildPlan.DiffArtifacts(oldBuildPlan, false)
+func (c *CveReport) Report(newCommit *buildplanner.Commit, oldCommit *buildplanner.Commit, names ...string) error {
+	changeset := newCommit.BuildPlan().DiffArtifacts(oldCommit.BuildPlan(), false)
 	if c.shouldSkipReporting(changeset) {
 		logging.Debug("Skipping CVE reporting")
 		return nil
@@ -72,8 +74,10 @@ func (c *CveReport) Report(newBuildPlan *buildplan.BuildPlan, oldBuildPlan *buil
 	}
 
 	if len(names) == 0 {
-		for _, ing := range ingredients {
-			names = append(names, ing.Name)
+		var err error
+		names, err = addedRequirements(oldCommit.BuildScript(), newCommit.BuildScript())
+		if err != nil {
+			return errs.Wrap(err, "Failed to get added requirements")
 		}
 	}
 	pg := output.StartSpinner(c.prime.Output(), locale.Tr("progress_cve_search", strings.Join(names, ", ")), constants.TerminalAnimationInterval)
@@ -198,4 +202,46 @@ func (c *CveReport) promptForSecurity() (bool, error) {
 	}
 
 	return confirm, nil
+}
+
+func addedRequirements(oldBuildScript *buildscript.BuildScript, newBuildScript *buildscript.BuildScript) ([]string, error) {
+	var names []string
+
+	old, err := oldBuildScript.Requirements()
+	if err != nil {
+		return nil, errs.Wrap(err, "Could not get old requirements")
+	}
+
+	oldReqs := make(map[string]bool)
+	for _, req := range old {
+		req, ok := req.(buildscript.DependencyRequirement)
+		if !ok {
+			continue
+		}
+		oldReqs[qualifiedName(req)] = true
+	}
+
+	newReqs, err := newBuildScript.Requirements()
+	if err != nil {
+		return nil, errs.Wrap(err, "Could not get new requirements")
+	}
+
+	for _, req := range newReqs {
+		req, ok := req.(buildscript.DependencyRequirement)
+		if !ok {
+			continue
+		}
+		if !oldReqs[qualifiedName(req)] {
+			names = append(names, req.Name)
+		}
+	}
+
+	return names, nil
+}
+
+func qualifiedName(req buildscript.DependencyRequirement) string {
+	if req.Namespace == "" {
+		return req.Name
+	}
+	return req.Namespace + "/" + req.Name
 }
